Refactored Datanode table to a separate component
