+++ hadoop-ozone/recon/src/main/resources/webapps/recon/ozone-recon-web/src/v2/pages/datanodes/datanodes.tsx
@@ -24,32 +24,20 @@ import React, {
 import moment from 'moment';
 import { AxiosError } from 'axios';
 import {
-  Table,
-  Tooltip,
-  Popover,
   Button,
   Modal
 } from 'antd';
 import {
-  ColumnsType,
-  TablePaginationConfig
-} from 'antd/es/table';
-import {
-  CheckCircleFilled,
-  CloseCircleFilled,
   DeleteOutlined,
-  HourglassFilled,
-  InfoCircleOutlined,
   WarningFilled,
 } from '@ant-design/icons';
 import { ValueType } from 'react-select';
 
 import Search from '@/v2/components/search/search';
-import StorageBar from '@/v2/components/storageBar/storageBar';
 import MultiSelect, { Option } from '@/v2/components/select/multiSelect';
+import DatanodesTable, { COLUMNS } from '@/v2/components/tables/datanodesTable';
 import AutoReloadPanel from '@/components/autoReloadPanel/autoReloadPanel';
 import { showDataFetchError } from '@/utils/common';
-import { ReplicationIcon } from '@/utils/themeIcons';
 import { AutoReloadHelper } from '@/utils/autoReloadHelper';
 import {
   AxiosGetHelper,
@@ -61,239 +49,13 @@ import { useDebounce } from '@/v2/hooks/debounce.hook';
 import {
   Datanode,
   DatanodeDecomissionInfo,
-  DatanodeOpState,
-  DatanodeOpStateList,
   DatanodeResponse,
   DatanodesResponse,
-  DatanodesState,
-  DatanodeState,
-  DatanodeStateList
+  DatanodesState
 } from '@/v2/types/datanode.types';
-import { Pipeline } from '@/v2/types/pipelines.types';
 
 import './datanodes.less'
-import DecommissionSummary from '@/v2/components/decommissioningSummary/decommissioningSummary';
-import { ColumnTitleProps } from 'antd/lib/table/interface';
-import { TableRowSelection } from 'antd/es/table/interface';
 
-moment.updateLocale('en', {
-  relativeTime: {
-    past: '%s ago',
-    s: '%ds',
-    m: '1min',
-    mm: '%dmins',
-    h: '1hr',
-    hh: '%dhrs',
-    d: '1d',
-    dd: '%dd',
-    M: '1m',
-    MM: '%dm',
-    y: '1y',
-    yy: '%dy'
-  }
-});
-
-const headerIconStyles: React.CSSProperties = {
-  display: 'flex',
-  alignItems: 'center'
-}
-
-const renderDatanodeState = (state: DatanodeState) => {
-  const stateIconMap = {
-    HEALTHY: <CheckCircleFilled twoToneColor='#1da57a' className='icon-success' />,
-    STALE: <HourglassFilled className='icon-warning' />,
-    DEAD: <CloseCircleFilled className='icon-failure' />
-  };
-  const icon = state in stateIconMap ? stateIconMap[state] : '';
-  return <span>{icon} {state}</span>;
-};
-
-const renderDatanodeOpState = (opState: DatanodeOpState) => {
-  const opStateIconMap = {
-    IN_SERVICE: <CheckCircleFilled twoToneColor='#1da57a' className='icon-success' />,
-    DECOMMISSIONING: <HourglassFilled className='icon-warning' />,
-    DECOMMISSIONED: <WarningFilled className='icon-warning' />,
-    ENTERING_MAINTENANCE: <HourglassFilled className='icon-warning' />,
-    IN_MAINTENANCE: <WarningFilled className='icon-warning' />
-  };
-  const icon = opState in opStateIconMap ? opStateIconMap[opState] : '';
-  return <span>{icon} {opState}</span>;
-};
-
-const getTimeDiffFromTimestamp = (timestamp: number): string => {
-  const timestampDate = new Date(timestamp);
-  return moment(timestampDate).fromNow();
-}
-
-const COLUMNS: ColumnsType<Datanode> = [
-  {
-    title: 'Hostname',
-    dataIndex: 'hostname',
-    key: 'hostname',
-    sorter: (a: Datanode, b: Datanode) => a.hostname.localeCompare(
-      b.hostname, undefined, { numeric: true }
-    ),
-    defaultSortOrder: 'ascend' as const
-  },
-  {
-    title: 'State',
-    dataIndex: 'state',
-    key: 'state',
-    filterMultiple: true,
-    filters: DatanodeStateList.map(state => ({ text: state, value: state })),
-    onFilter: (value, record: Datanode) => record.state === value,
-    render: (text: DatanodeState) => renderDatanodeState(text),
-    sorter: (a: Datanode, b: Datanode) => a.state.localeCompare(b.state)
-  },
-  {
-    title: 'Operational State',
-    dataIndex: 'opState',
-    key: 'opState',
-    filterMultiple: true,
-    filters: DatanodeOpStateList.map(state => ({ text: state, value: state })),
-    onFilter: (value, record: Datanode) => record.opState === value,
-    render: (text: DatanodeOpState) => renderDatanodeOpState(text),
-    sorter: (a: Datanode, b: Datanode) => a.opState.localeCompare(b.opState)
-  },
-  {
-    title: 'UUID',
-    dataIndex: 'uuid',
-    key: 'uuid',
-    sorter: (a: Datanode, b: Datanode) => a.uuid.localeCompare(b.uuid),
-    defaultSortOrder: 'ascend' as const,
-    render: (uuid: string, record: Datanode) => {
-      return (
-        //1. Compare Decommission Api's UUID with all UUID in table and show Decommission Summary
-        (decommissionUuids && decommissionUuids.includes(record.uuid) && record.opState !== 'DECOMMISSIONED') ?
-          <DecommissionSummary uuid={uuid} /> : <span>{uuid}</span>
-      );
-    }
-  },
-  {
-    title: 'Storage Capacity',
-    dataIndex: 'storageUsed',
-    key: 'storageUsed',
-    sorter: (a: Datanode, b: Datanode) => a.storageRemaining - b.storageRemaining,
-    render: (_: string, record: Datanode) => (
-      <StorageBar
-        strokeWidth={6}
-        capacity={record.storageTotal}
-        used={record.storageUsed}
-        remaining={record.storageRemaining}
-        committed={record.storageCommitted} />
-    )
-  },
-  {
-    title: 'Last Heartbeat',
-    dataIndex: 'lastHeartbeat',
-    key: 'lastHeartbeat',
-    sorter: (a: Datanode, b: Datanode) => moment(a.lastHeartbeat).unix() - moment(b.lastHeartbeat).unix(),
-    render: (heartbeat: number) => {
-      return heartbeat > 0 ? getTimeDiffFromTimestamp(heartbeat) : 'NA';
-    }
-  },
-  {
-    title: 'Pipeline ID(s)',
-    dataIndex: 'pipelines',
-    key: 'pipelines',
-    render: (pipelines: Pipeline[], record: Datanode) => {
-      const renderPipelineIds = (pipelineIds: Pipeline[]) => {
-        return pipelineIds?.map((pipeline: any, index: any) => (
-          <div key={index} className='pipeline-container-v2'>
-            <ReplicationIcon
-              replicationFactor={pipeline.replicationFactor}
-              replicationType={pipeline.replicationType}
-              leaderNode={pipeline.leaderNode}
-              isLeader={pipeline.leaderNode === record.hostname} />
-            {pipeline.pipelineID}
-          </div >
-        ))
-      }
-
-      return (
-        <Popover
-          content={
-            renderPipelineIds(pipelines)
-          }
-          title="Related Pipelines"
-          placement="bottomLeft"
-          trigger="hover">
-          <strong>{pipelines.length}</strong> pipelines
-        </Popover>
-      );
-    }
-  },
-  {
-    title: () => (
-      <span style={headerIconStyles} >
-        Leader Count
-        <Tooltip
-          title='The number of Ratis Pipelines in which the given datanode is elected as a leader.' >
-          <InfoCircleOutlined style={{ paddingLeft: '4px' }} />
-        </Tooltip>
-      </span>
-    ),
-    dataIndex: 'leaderCount',
-    key: 'leaderCount',
-    sorter: (a: Datanode, b: Datanode) => a.leaderCount - b.leaderCount
-  },
-  {
-    title: 'Containers',
-    dataIndex: 'containers',
-    key: 'containers',
-    sorter: (a: Datanode, b: Datanode) => a.containers - b.containers
-  },
-  {
-    title: () => (
-      <span style={headerIconStyles}>
-        Open Container
-        <Tooltip title='The number of open containers per pipeline.'>
-          <InfoCircleOutlined style={{ paddingLeft: '4px' }} />
-        </Tooltip>
-      </span>
-    ),
-    dataIndex: 'openContainers',
-    key: 'openContainers',
-    sorter: (a: Datanode, b: Datanode) => a.openContainers - b.openContainers
-  },
-  {
-    title: 'Version',
-    dataIndex: 'version',
-    key: 'version',
-    sorter: (a: Datanode, b: Datanode) => a.version.localeCompare(b.version),
-    defaultSortOrder: 'ascend' as const
-  },
-  {
-    title: 'Setup Time',
-    dataIndex: 'setupTime',
-    key: 'setupTime',
-    sorter: (a: Datanode, b: Datanode) => a.setupTime - b.setupTime,
-    render: (uptime: number) => {
-      return uptime > 0 ? moment(uptime).format('ll LTS') : 'NA';
-    }
-  },
-  {
-    title: 'Revision',
-    dataIndex: 'revision',
-    key: 'revision',
-    sorter: (a: Datanode, b: Datanode) => a.revision.localeCompare(b.revision),
-    defaultSortOrder: 'ascend' as const
-  },
-  {
-    title: 'Build Date',
-    dataIndex: 'buildDate',
-    key: 'buildDate',
-    sorter: (a: Datanode, b: Datanode) => a.buildDate.localeCompare(b.buildDate),
-    defaultSortOrder: 'ascend' as const
-  },
-  {
-    title: 'Network Location',
-    dataIndex: 'networkLocation',
-    key: 'networkLocation',
-    sorter: (a: Datanode, b: Datanode) => a.networkLocation.localeCompare(b.networkLocation),
-    defaultSortOrder: 'ascend' as const
-  }
-];
 
 const defaultColumns = COLUMNS.map(column => ({
   label: (typeof column.title === 'string')
@@ -340,19 +102,6 @@ const Datanodes: React.FC<{}> = () => {
 
   function handleColumnChange(selected: ValueType<Option, true>) {
     setSelectedColumns(selected as Option[]);
-  }
-
-  function filterSelectedColumns() {
-    const columnKeys = selectedColumns.map((column) => column.value);
-    return COLUMNS.filter(
-      (column) => columnKeys.indexOf(column.key as string) >= 0
-    );
-  }
-
-  function getFilteredData(data: Datanode[]) {
-    return data.filter(
-      (datanode: Datanode) => datanode[searchColumn].includes(debouncedSearch)
-    );
   }
 
   async function loadDecommisionAPI() {
@@ -462,9 +211,8 @@ const Datanodes: React.FC<{}> = () => {
     });
   }, []);
 
-  function isSelectable(record: Datanode) {
-    // Disable checkbox for any datanode which is not DEAD to prevent removal
-    return record.state !== 'DEAD' && true;
+  function handleSelectionChange(rows: React.Key[]) {
+    setSelectedRows(rows);
   }
 
   function handleModalOk() {
@@ -478,21 +226,6 @@ const Datanodes: React.FC<{}> = () => {
   };
 
   const { dataSource, lastUpdated, columnOptions } = state;
-
-  const rowSelection: TableRowSelection<Datanode> = {
-    selectedRowKeys: selectedRows,
-    onChange: (rows: React.Key[]) => { setSelectedRows(rows) },
-    getCheckboxProps: (record: Datanode) => ({
-      disabled: isSelectable(record)
-    }),
-  };
-
-  const paginationConfig: TablePaginationConfig = {
-    showTotal: (total: number, range) => (
-      `${range[0]}-${range[1]} of ${total} Datanodes`
-    ),
-    showSizeChanger: true
-  };
 
   return (
     <>
@@ -543,17 +276,15 @@ const Datanodes: React.FC<{}> = () => {
                 setSearchColumn(value as 'hostname' | 'uuid' | 'version' | 'revision')
               }} />
           </div>
-          <div>
-            <Table
-              rowSelection={rowSelection}
-              dataSource={getFilteredData(dataSource)}
-              columns={filterSelectedColumns()}
-              loading={loading}
-              rowKey='uuid'
-              pagination={paginationConfig}
-              scroll={{ x: 'max-content', scrollToFirstRowOnChange: true }}
-              locale={{ filterTitle: '' }} />
-          </div>
+          <DatanodesTable
+            loading={loading}
+            data={dataSource}
+            selectedColumns={selectedColumns}
+            selectedRows={selectedRows}
+            searchColumn={searchColumn}
+            searchTerm={debouncedSearch}
+            handleSelectionChange={handleSelectionChange}
+            decommissionUuids={decommissionUuids} />
         </div>
       </div>
       <Modal
